requirements.txt
pandas
redis
openai
streamlit
streamlit-chat
python-dotenv
